refactor: exclude option is now an array of values
Having exclude as an object presented some odd ways to describe which parts of the state you wanted
to exclude. The array format is more expressive and makes general cases where the entire reducer is
targeted for exclusion simpler and avoids the use of a special character

The type of exclude has changed from object to array. Refer to the documentation to see how to
supply values to it now.